refactor(meeting): lift SessionManager lifecycle methods into peer module (closes #488)
Pure code-move refactor following the audit's primary recommendation
for #431 item 2 (which #487 deferred).

`start_manual` (~258 LOC), `stop_manual` (~89 LOC), and
`append_if_active` (~38 LOC) move into a new `meeting/lifecycle.rs`
peer. The state machine + struct definitions stay in `manager.rs`
along with `new`, `new_for_test`, `reconcile_orphan_sessions`,
`current_partials_for`, `active_session_id`, and the `Drop` impl
— each file now has one job.

`SessionManager` fields, `ActiveSession`, and `SessionState` are
now `pub(super)` (visible only within `crate::meeting::*`) so the
lifecycle peer can drive the state machine without going through
accessor noise. The widening is scoped: nothing outside `meeting`
sees the new visibility.

manager.rs drops from 2,008 → 1,593 LOC. lifecycle.rs lands at 474.

Locking discipline preserved: the state mutex is never held across
an `.await`. See the lifecycle module's docs + `learnings.md`
2026-05-02 for the rationale.
